+++ backend/packages/caching/CHANGELOG.md
@@ -1,5 +1,19 @@
 # @cat-factory/caching
 
+## 0.20.63
+
+### Patch Changes
+
+- Updated dependencies [9dfd40b]
+  - @cat-factory/kernel@0.331.0
+
+## 0.20.62
+
+### Patch Changes
+
+- Updated dependencies [1c79070]
+  - @cat-factory/kernel@0.330.0
+
 ## 0.20.61
 
 ### Patch Changes